Small fix to Nguyen
git-svn-id: http://encog-java.googlecode.com/svn/trunk/encog-core@1522 f90f6e9a-ac51-0410-b353-d1b83c6f6923

+++ src/org/encog/mathutil/randomize/NguyenWidrowRandomizer.java
@@ -77,7 +77,7 @@ public class NguyenWidrowRandomizer extends RangeRandomizer implements Randomize
                 
           while ( !previousLayer.equals( network.getLayer( BasicNetwork.TAG_OUTPUT ) )&& iterator.hasNext() ) {
               final Layer     current         = iterator.next();
-              final double    beta            =  0.7d * Math.pow( current.getNeuronCount() , ( 1/numOfInputNeurons ) );
+              final double    beta            =  0.7d * Math.pow( (double)current.getNeuronCount() , (double)( 1.0/numOfInputNeurons ) );
               final double    normOfWeight[]  = getNormOfWeight(  previousLayer.getNext() );
               
               for (final Synapse synapse : previousLayer.getNext()) {
